requirements.txt
fastapi[all]
uvicorn
python-multipart
google-generativeai
pillow